--- apps/server/api/stripe/methods/_methodId@string/index.ts
@@ -1,10 +1,0 @@
-import type { AuthHeader } from '@frourio-demo/types';
-
-export type Methods = {
-  put: {
-    reqHeaders: AuthHeader;
-  };
-  delete: {
-    reqHeaders: AuthHeader;
-  };
-}
