+++ apps/server/api/stripe/methods/controller.ts
@@ -1,6 +1,12 @@
 import { defineController } from './$relay';
-import { getPaymentMethods } from '$/domains/stripe';
+import { getPaymentMethods, attachPaymentMethod, detachPaymentMethod } from '$/domains/stripe';
 
-export default defineController(({ getPaymentMethods }), ({ getPaymentMethods }) => ({
+export default defineController(({ getPaymentMethods, attachPaymentMethod, detachPaymentMethod }), ({
+  getPaymentMethods,
+  attachPaymentMethod,
+  detachPaymentMethod,
+}) => ({
   get: async({ user }) => getPaymentMethods(user.id),
+  put: async({ user, body }) => attachPaymentMethod(body.methodId, user.id),
+  delete: async({ body }) => detachPaymentMethod(body.methodId),
 }));
