+++ apps/server/api/stripe/methods/index.ts
@@ -5,5 +5,13 @@ export type Methods = {
   get: {
     reqHeaders: AuthHeader;
     resBody: Stripe.PaymentMethod[]
-  }
+  };
+  put: {
+    reqHeaders: AuthHeader;
+    reqBody: { methodId: string }
+  };
+  delete: {
+    reqHeaders: AuthHeader;
+    reqBody: { methodId: string }
+  };
 }
